Rename and clean up
