--- src/components/scalingSVG/scalingSVG.module.scss
@@ -1,18 +1,0 @@
-@import '~/scss/variables.scss';
-
-.scalingSvgContainer {
-  position: relative;
-  height: 0;
-  width: 100%;
-  padding: 0;
-  padding-bottom: 100%;
-  margin: $default-margin 0;
-
-  svg {
-    position: absolute;
-    height: 100%;
-    width: 100%;
-    left: 0;
-    top: 0;
-  }
-}
