--- src/components/titleBlock/titleBlock.module.scss
@@ -1,40 +1,0 @@
-@import '~/scss/variables.scss';
-
-.titleRow {
-  display: flex;
-  align-items: flex-start;
-  margin-bottom: $default-margin;
-
-  svg {
-    fill: currentColor;
-    color: $icon-color;
-    width: 3rem;
-    height: auto;
-    max-height: 3rem;
-    margin-right: $default-margin;
-  }
-
-  h2 {
-    margin: 0;
-  }
-
-  p {
-    max-width: 20rem;
-    margin-top: $default-margin / 2;
-  }
-
-  @media screen and (min-width: 768px) {
-    margin-bottom: $default-margin * 2;
-
-    svg {
-      width: 4rem;
-      height: auto;
-      max-height: 4rem;
-      margin-right: $default-margin * 2;
-    }
-
-    p {
-      margin-top: $default-margin;
-    }
-  }
-}
