+++ src/domain/layout/components/safety-region-combo-box.tsx
