+++ src/pages/gemeente/[code]/positief-geteste-mensen.tsx
@@ -13,7 +13,7 @@ import { MunicipalityChoropleth } from '~/components/choropleth/municipality-cho
 import { createSelectMunicipalHandler } from '~/components/choropleth/select-handlers/create-select-municipal-handler';
 import { createPositiveTestedPeopleMunicipalTooltip } from '~/components/choropleth/tooltips/municipal/create-positive-tested-people-municipal-tooltip';
 import { FCWithLayout } from '~/domain/layout/layout';
-import { getMunicipalityLayout } from '~/domain/layout/MunicipalityLayout';
+import { getMunicipalityLayout } from '~/domain/layout/municipality-layout';
 import { SEOHead } from '~/components/seoHead';
 import {
   getMunicipalityData,
