+++ src/pages/veiligheidsregio/index.tsx
@@ -13,8 +13,8 @@ import { createSelectRegionHandler } from '~/components/choropleth/select-handle
 import { escalationTooltip } from '~/components/choropleth/tooltips/region/escalation-tooltip';
 import styles from '~/components/choropleth/tooltips/tooltip.module.scss';
 import { FCWithLayout } from '~/domain/layout/layout';
-import { SafetyRegionComboBox } from '~/domain/layout/safety-region-combo-box';
-import { getSafetyRegionLayout } from '~/domain/layout/SafetyRegionLayout';
+import { SafetyRegionComboBox } from '~/domain/layout/components/safety-region-combo-box';
+import { getSafetyRegionLayout } from '~/domain/layout/safety-region-layout';
 import { SEOHead } from '~/components/seoHead';
 import { TALLLanguages } from '~/locale/index';
 import { parseMarkdownInLocale } from '~/utils/parse-markdown-in-locale';
